+++ app/services/movie_service.rb
@@ -1,4 +1,4 @@
-class MovieService < Service
+class MovieService < APIService
   def self.get_top_rated_movies(page)
     response = conn.get('/3/movie/top_rated') do |f|
       f.params['page'] = page
@@ -12,22 +12,14 @@ class MovieService < Service
   end
 
   def self.search_movie_by_title(title)
-    response = conn.get("/3/search/movie?query=#{title}")
+    response = conn.get("/3/search/movie") do |f|
+      f.params['query'] = title
+    end
     parse_json(response)
   end
 
   def self.movie_info(movie_id)
-    response = Faraday.get(URL + movie_id, { api_key: ENV['movie_key'] })
-    JSON.parse(response.body, symbolize_names: true)
-  end
-
-  def self.conn
-    Faraday.new('https://api.themoviedb.org') do |f|
-      f.params['api_key'] = ENV['movie_key']
-    end
-  end
-
-  def self.parse_json(response)
-    JSON.parse(response.body, symbolize_names: true)
+    response = conn.get(URL + movie_id)
+    parse_json(response)
   end
 end
